backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Task queue
celery[redis]==5.4.0
redis==5.2.1

# Validation
pydantic==2.10.4
pydantic-settings==2.7.1

# LLM clients
openai==1.59.7
anthropic==0.42.0
google-genai==1.0.0

# SERP
serpapi==0.1.5

# Utilities
httpx==0.28.1
python-dotenv==1.0.1
tenacity==9.0.0
beautifulsoup4==4.12.3
